Superadmin

+++ src/user/services/role/role.service.ts
@@ -32,6 +32,11 @@ export class RoleService
     private readonly request: Request & Express.Request,
   ) {}
 
+  private checkSuperAdmin(role: Role) {
+    if (role.superAdmin === false) return;
+    throw new ForbiddenException('Cannot modify superadmin role or user');
+  }
+
   private checkAuthenticatedUserPermissions(
     role: RoleDto,
     message = 'Current user is missing required permissions',
@@ -75,6 +80,7 @@ export class RoleService
     const target: Role = {
       ...entity,
     };
+    this.checkSuperAdmin(target);
     return await this.roleRepo.save(target).catch(() => {
       throw new BadRequestException();
     });
@@ -82,12 +88,9 @@ export class RoleService
 
   async update(id: string | number, entity: UpdateRoleDto): Promise<void> {
     this.checkAuthenticatedUserPermissions(entity);
-    let role: Role;
-    try {
-      role = await this.roleRepo.findOneOrFail({ where: { id: id } });
-    } catch {
-      throw new NotFoundException(`Role does not exist at id : ${id}`);
-    }
+    const role = await this.find(id);
+    this.checkSuperAdmin(role);
+    this.checkSuperAdmin(entity);
 
     this.checkAuthenticatedUserPermissions(role);
 
@@ -96,11 +99,13 @@ export class RoleService
       name: entity.name,
       permissions: entity.permissions,
     };
+    this.checkSuperAdmin(target);
 
     await this.roleRepo.save(target);
   }
 
   async deleteFromId(id: string | number): Promise<void> {
+    this.checkSuperAdmin(await this.find(id));
     const result = await this.roleRepo.delete(id);
     if (result.affected < 1) {
       throw new BadRequestException('Role not found or already deleted');
@@ -108,6 +113,7 @@ export class RoleService
   }
 
   async delete(entity: Role): Promise<void> {
+    this.checkSuperAdmin(entity);
     const result = await this.roleRepo.delete(entity);
     if (result.affected < 1) {
       throw new BadRequestException('Role not found or already deleted');
